move internal definitions of http2 impl. to http2_internal.h (this effectively hides khash.h from public headers)
